Add Earl::Logger extension module
Replaces Crystal's Logger with a custom implementation, using a
global supervised Agent that is sent messages to log through a
Mailbox.

Adds an additional VERBOSE severity level between DEBUG and INFO,
and a SILENT level to disable any log.

Adds support for a backend to customize where logs are sent to. A
default ConsoleBackend is available to write a line to any IO.
Earl::Logger will write to STDERR by default. Custom backends may be
created by extending Earl::Logger::Backend, for example to target a
syslog server.