Introduce `Earl.application`
`Earl.application` is a singleton `Earl::Supervisor` meant to hold
the main actors of an application. For example `Earl::Logger` is now
handled by it.

Unlike simple supervisors, `Earl.application` will gracefully stop
when `SIGINT` or `SIGTERM` is received or the program called `exit`.

BREAKING CHANGE: programs are expected to start `Earl.application`,
at least to start the logger actor, otherwise the application will
hang once the logger mailbox is full. It may be spawned and
forgotten, but we advise to leverage it as the main supervisor of
your program, since it will block the main fiber from terminating
up until the program it told to terminate.