Fix: flush pending logger messages
Messages could be lost while the program was exiting because the
logger actor didn't try to flush its mailbox.

Some messages may still be lost if a spawned and forgotten actor
still logs messages, but the logger actor should be stopped last
from `Earl.application` and other actors should have stopped, so
they hopefully shall be kept to a minimum.