Introduce socket server agents (TCP, SSL, UNIX)
Unlike other agents that usually invoke a `#call` method, these
servers take a block that will be invoked with the incoming socket.

- `Earl::TCPServer` wraps a `TCPServer`.
- `Earl::SSLServer` wraps a `SSLServer` and also wraps incoming
  connections with `OpenSSL::SSL::Socket::Server` to achieve
  transparent SSL servers.
- `Earl::UNIXServer` wraps a `UNIXServer`.

These agents may be used directly, but are designed for the
`Earl::SockServer` agent that permits to start and supervise as
many servers as required, and will invoke its `#call(socket)`
method.