Fix: supervisor shall stop running agents only
There is a time window between the moment a supervised agent is
recycled and the moment it will be restarted. During this lapse of
time, the supervisor may be told to stop, which in turn will tell
all supervised agents to stop, regardless of their current state,
which led to transition errors (e.g. starting -> stopping).

Furthermore, recycled agents were always restarted, regardless of
the current supervisor state. They should only be restarted if the
supervisor itself is running, and merely return otherwise.