Fix: abstract def call(M) must be implemented
If M is an union type, crystal now requires the exact call(M) metho
to exist, and doesn't consider overloads. This patch changes the
abstract method for a method that raises in a macro with some
macro-fu to try to report a helpful message to the developer,
telling her which types are missing.