Initial concurrency primitives
Implements the following, common, concurrency primitives to handle
concurrency and parallelism safety:

- Lock: a module to avoid duplicating lock related functions;
- SpinLock: a multithread lock mechanism using CPU spin/pause instead of
  thread/fiber context switches, strictly meant to protect access to
  thread shared data structures;
- WaitList: keeps a singly list of fibers waiting to be resumed;
- UnsafeMutex: fiber-aware mutually exclusive lock (unchecked);
- ConditionVariable: fiber synchronization;
- WaitGroup: wait until other fibers are done;
- RWLock: a fiber-aware many-readers single-write lock.

Those are `struct` to be embedded right into the objects that need them
and not allocated through the GC, improving the locality. For example
RWLock relies on both ConditionVariable and UnsafeMutex.

Being structs mean that they must be passed around as pointers or
accessed directly (as a local/instance variable) otherwise they'd be
copied and thus invalid!