Drop `Earl::Mailbox(A, M)`
It wasn't implemented as per the spec as an alternative fanout module to
`Earl::Mailbox(M)` but as a distinct sync class. It's also using a named
that would better fit a registry of actors, so we could reference by
names instead of having to pass an explicit reference to the object
around.

closes #8
closes #27